# perf: faster prefix record deser