--- a/crates/rattler_conda_types/src/prefix_record.rs
+++ b/crates/rattler_conda_types/src/prefix_record.rs
@@ -267,9 +267,7 @@ impl PrefixRecord {
     }
 
     fn from_str_mut(s: &mut str) -> Result<Self, std::io::Error> {
-        // let mut d = s.to_vec();
         unsafe { simd_json::serde::from_slice(s.as_bytes_mut()).map_err(Into::into) }
-        // serde_json::from_str(s).map_err(Into::into)
     }
 }
 
@@ -277,10 +275,6 @@ impl FromStr for PrefixRecord {
     type Err = std::io::Error;
 
     fn from_str(s: &str) -> Result<Self, Self::Err> {
-        // let mut d = s.to_vec();
-        // unsafe {
-        //     simd_json::serde::from_slice(s.as_bytes_mut()).map_err(Into::into)
-        // }
         serde_json::from_str(s).map_err(Into::into)
     }
 }