--- a/crates/rattler_conda_types/src/prefix_record.rs
+++ b/crates/rattler_conda_types/src/prefix_record.rs
@@ -267,6 +267,10 @@ impl PrefixRecord {
     }
 
     fn from_str_mut(s: &mut str) -> Result<Self, std::io::Error> {
+        // here we are using unsafe block for as_bytes_mut which has the same safety guarantees as str::as_bytes_mut
+        // quoting: "the caller must ensure that the content of the slice is valid UTF-8
+        // before the borrow ends and the underlying `str` is used."
+        // In our case, underlying `str` is not used after the borrow ends.
         unsafe { simd_json::serde::from_slice(s.as_bytes_mut()).map_err(Into::into) }
     }
 }