… py-rattler/Cargo.lock
… test-data/conda-meta/icu-73.2-hc8870d7_0.json
… test-data/conda-meta/libcblas-3.9.0-20_osxarm64_openblas.json
… test-data/conda-meta/libclang-16.0.6-default_he012953_6.json
… test-data/conda-meta/libclang13-16.0.6-default_h83d0a53_6.json
… test-data/conda-meta/libcxx-17.0.6-h5f092b4_0.json
… test-data/conda-meta/libexpat-2.6.2-hebf3989_0.json
… test-data/conda-meta/libffi-3.4.2-h3422bc3_5.json
… test-data/conda-meta/libgfortran-5.0.0-13_2_0_hd922786_3.json
… test-data/conda-meta/libgfortran5-13.2.0-hf226fd6_3.json
… test-data/conda-meta/libiconv-1.17-h0d3ecfb_2.json
… test-data/conda-meta/liblapack-3.9.0-20_osxarm64_openblas.json
… test-data/conda-meta/libllvm16-16.0.6-haab561b_3.json
… test-data/conda-meta/libopenblas-0.3.25-openmp_h6c19121_0.json
… test-data/conda-meta/libsqlite-3.45.3-h091b4b1_0.json
… test-data/conda-meta/libxml2-2.12.7-ha661575_0.json
… test-data/conda-meta/libzlib-1.2.13-h53f4e23_5.json
… test-data/conda-meta/llvm-openmp-18.1.6-hde57baf_0.json
… test-data/conda-meta/ncurses-6.5-hb89a1cb_0.json
… test-data/conda-meta/numpy-1.26.4-py311h7125741_0.json
… test-data/conda-meta/openssl-3.3.0-hfb2fe0b_3.json
… test-data/conda-meta/python-3.11.9-h932a869_0_cpython.json
… test-data/conda-meta/python_abi-3.11-4_cp311.json
… test-data/conda-meta/readline-8.2-h92ec313_1.json
… test-data/conda-meta/scipy-1.11.4-py311h2b215a9_0.json
… test-data/conda-meta/tk-8.6.13-h5083fa2_1.json
… test-data/conda-meta/tzdata-2024a-h0c530f3_0.json
… test-data/conda-meta/xz-5.2.6-h57fd34a_0.json
… test-data/conda-meta/zstd-1.5.6-hb46c0d2_0.json