based on this blog https://blog.datalust.co/deserializing-json-really-fast/ I've tried to choose simd_json for our deserialization to PrefixRecord as it seems that it has a lot of samples in the flamegraph
![image](https://github.com/mamba-org/rattler/assets/25200469/72f8172d-a28f-4458-8a76-f45032e0dc32)

In the benchmarks, having 42 JSON files in the conda_meta folder and using simd_json will give us the following results:

```
process_json_files      time:   [9.2815 ms 9.3741 ms 9.4924 ms]
Found 7 outliers among 100 measurements (7.00%)
5 (5.00%) high mild
2 (2.00%) high severe
```


using old way:
```
process_json_files      time:   [9.7586 ms 9.8881 ms 10.041 ms]
change: [+3.5487% +5.4835% +7.3793%] (p = 0.00 < 0.05)
Performance has regressed.
```

We can see an increase of `+5.4835%`