@Hofer-Julian 2024-07-12
I pushed a commit that removes the unsafe usage: https://github.com/mamba-org/rattler/pull/777/commits/11c86003dc3834fe5570734d115ff2d70e4394a1